study-mate/studyio/requirements.txt
Flask==2.3.3
Flask-Cors==4.0.0
Pillow==10.0.0
google-cloud-vision==3.4.0
openai==1.1.1
python-dotenv==1.0.0
gunicorn==21.2.0
